Separate types into multiple files
Other Changes:
- Use enums for eviction config fields
- Use enum types in eviction config decoder
- Add isDefined field into HotRestartConfig, MerkleTreeConfig and EventJournalConfig
