+++ controllers/protocol/types/attribute_config.go
@@ -1,0 +1,6 @@
+package types
+
+type AttributeConfig struct {
+	Name               string
+	ExtractorClassName string
+}
