+++ controllers/protocol/types/bitmap_index_options.go
@@ -1,0 +1,6 @@
+package types
+
+type BitmapIndexOptions struct {
+	UniqueKey               string
+	UniqueKeyTransformation UniqueKeyTransformation
+}
