+++ controllers/protocol/types/eviction_config_holder.go
@@ -1,0 +1,11 @@
+package types
+
+import iserialization "github.com/hazelcast/hazelcast-go-client"
+
+type EvictionConfigHolder struct {
+	Size                int32
+	MaxSizePolicy       MaxSizePolicyType
+	EvictionPolicy      EvictionPolicyType
+	ComparatorClassName string
+	Comparator          iserialization.Data
+}
