+++ controllers/protocol/types/index_config.go
@@ -1,0 +1,8 @@
+package types
+
+type IndexConfig struct {
+	Name               string
+	Type               IndexType
+	Attributes         []string
+	BitmapIndexOptions BitmapIndexOptions
+}
