+++ controllers/protocol/types/listener_config_holder.go
@@ -1,0 +1,11 @@
+package types
+
+import iserialization "github.com/hazelcast/hazelcast-go-client"
+
+type ListenerConfigHolder struct {
+	ListenerType           int32
+	ListenerImplementation iserialization.Data
+	ClassName              string
+	IncludeValue           bool
+	Local                  bool
+}
