+++ controllers/protocol/types/map_store_config_holder.go
@@ -1,0 +1,16 @@
+package types
+
+import iserialization "github.com/hazelcast/hazelcast-go-client"
+
+type MapStoreConfigHolder struct {
+	Enabled               bool
+	WriteCoalescing       bool
+	WriteDelaySeconds     int32
+	WriteBatchSize        int32
+	ClassName             string
+	Implementation        iserialization.Data
+	FactoryClassName      string
+	FactoryImplementation iserialization.Data
+	Properties            map[string]string
+	InitialLoadMode       string
+}
