+++ controllers/protocol/types/near_cache_preloader_config.go
@@ -1,0 +1,8 @@
+package types
+
+type NearCachePreloaderConfig struct {
+	Enabled                  bool
+	Directory                string
+	StoreInitialDelaySeconds int32
+	StoreIntervalSeconds     int32
+}
