+++ controllers/protocol/types/predicate_config_holder.go
@@ -1,0 +1,9 @@
+package types
+
+import iserialization "github.com/hazelcast/hazelcast-go-client"
+
+type PredicateConfigHolder struct {
+	ClassName      string
+	Sql            string
+	Implementation iserialization.Data
+}
