+++ controllers/protocol/types/query_cache_config_holder.go
@@ -1,0 +1,16 @@
+package types
+
+type QueryCacheConfigHolder struct {
+	BatchSize             int32
+	BufferSize            int32
+	DelaySeconds          int32
+	IncludeValue          bool
+	Populate              bool
+	Coalesce              bool
+	InMemoryFormat        string
+	Name                  string
+	PredicateConfigHolder PredicateConfigHolder
+	EvictionConfigHolder  EvictionConfigHolder
+	ListenerConfigs       []ListenerConfigHolder
+	IndexConfigs          []IndexConfig
+}
