+++ controllers/protocol/types/wan_replication_ref.go
@@ -1,0 +1,8 @@
+package types
+
+type WanReplicationRef struct {
+	Name                 string
+	MergePolicyClassName string
+	Filters              []string
+	RepublishingEnabled  bool
+}
